Refactor: Extract nodes to modular files with shared utilities
- Extract router_node, verify_correctness_node, maintain_character_node, format_output_node to individual files
- Move shared utilities (classify_intent, OFF_TOPIC_RESPONSES) to langgraph_flow.py
- Update app.py imports to use modular structure
- Preserve all existing functionality and LangSmith tracing
- All tests pass, API integration confirmed
