+++ BACKLOG.md
@@ -4,13 +4,7 @@
 
 ### 🔧 Technical Debt & Architecture
 
-#### 1. Complete AgentExecutor → LangGraph Migration
-- **Priority**: Low
-- **Status**: Pending
-- **Description**: Remove vestigial AgentExecutor from `find_hint_node` for cleaner architecture
-- **Details**: Currently using hybrid approach - LangGraph nodes calling old AgentExecutor chain
-- **Files**: `src/agent.py`, `src/langgraph_flow.py`
-- **Benefit**: Cleaner code, faster execution, full LangGraph architecture
+
 
 #### 2. Externalize LLM Model References  
 - **Priority**: Low
@@ -65,13 +59,7 @@
 
 ### 📊 Evaluation & Debugging
 
-#### 4. Add RAG Pipeline Diagnostic Logging
-- **Priority**: Medium
-- **Status**: Pending
-- **Description**: Add intermediate logging to identify retrieval vs generation issues
-- **Details**: Log retrieved documents, context quality, and response generation steps
-- **Files**: `src/agent.py` (hint_lookup function)
-- **Benefit**: Quick debugging without full RAGAS evaluation, identify pipeline bottlenecks
+
 
 #### 5. Implement Full RAGAS Evaluation
 - **Priority**: Low
@@ -108,15 +96,20 @@
 - ✅ **Enhanced LangSmith tracing** - Clear input/output visibility for debugging
 - ✅ **Canned responses** - Efficient off-topic handling without extra LLM calls
 
+### Technical Architecture Improvements
+- ✅ **Complete AgentExecutor → LangGraph Migration** - Removed all vestigial AgentExecutor code for clean LangGraph architecture
+- ✅ **RAG Pipeline Diagnostic Logging** - Added comprehensive logging for retrieval and generation steps
+
 ---
 
 ## Notes
 
 - **Demo Day Target**: August 25, 2025
+- **Code Freeze Target**: August 21, 2025
 - **Current State**: Production ready with smart routing and character integrity
 - **Architecture**: LangGraph-based flow with temporary AgentExecutor bridge
 - **Focus**: Prioritize user-facing features over technical debt until post-demo
 
 ---
 
-*Last Updated: January 18, 2025*
+*Last Updated: Aug 20, 2025*
