+++ src/app.py
@@ -5,13 +5,11 @@ from dotenv import load_dotenv
 from langgraph.graph import StateGraph, START, END
 from state import LangGraphState
 from langsmith import traceable
-from langgraph_flow import (
-    router_node, 
-    find_hint_node, 
-    verify_correctness_node, 
-    maintain_character_node, 
-    format_output_node
-)
+from router_node import router_node
+from find_hint_node import find_hint_node  
+from verify_correctness_node import verify_correctness_node
+from maintain_character_node import maintain_character_node
+from format_output_node import format_output_node
 
 # Load environment variables
 load_dotenv()
