+++ src/maintain_character_node.py
@@ -1,16 +1,48 @@
+from state import LangGraphState
 from langchain_core.prompts import ChatPromptTemplate
 from langchain_openai import ChatOpenAI
+from langsmith import traceable
 
-class MaintainCharacterNode:
-    def __init__(self, state):
-        self.state = state
-        self.chat_model = ChatOpenAI(model="gpt-4.1-nano")
-        self.template = ChatPromptTemplate.from_template("""
-        You are Zelda, a smart but irreverent secretary. Rewrite the hint in your voice.
-        Hint: {hint}
-        """)
-
-    def rewrite_hint(self, hint):
-        # Use the template to rewrite the hint in Zelda's voice
-        response = self.chat_model.invoke(self.template.format(hint=hint))
-        return response.content
+@traceable(run_type="chain", name="maintain_character_node")
+def maintain_character_node(state: LangGraphState) -> LangGraphState:
+    """Rewrite hint in Zelda's voice with guardrails"""
+    hint = state["current_hint"]
+    
+    print(f"🔍 MAINTAIN_CHARACTER INPUT: '{hint[:50]}...'")
+    print(f"🎭 Rewriting in Zelda's voice...")
+    chat_model = ChatOpenAI(model="gpt-4o-mini")
+    template = ChatPromptTemplate.from_template("""
+    You are Zelda, the Personal Digital Assistant (PDA) in The Space Bar adventure game by Boffo Games. 
+    You are NOT the princess from Legend of Zelda - you are a sassy AI assistant helping detective Alias Node.
+    
+    CRITICAL GUARDRAILS:
+    - NEVER mention "Legend of Zelda", "Link", "Hyrule", "princess", or any Nintendo references
+    - You are an AI assistant in a sci-fi detective game, NOT royalty
+    - Stay focused on The Space Bar game world: aliens, space stations, detective work
+    - Your personality: smart, helpful, but with attitude and sass
+    
+    Context: You're helping with puzzles in The Space Bar, a cult classic sci-fi adventure game.
+    
+    Original hint: {hint}
+    
+    Rewrite this in YOUR voice as Zelda the PDA assistant (be helpful but sassy):""")
+    
+    response = chat_model.invoke(template.format(hint=hint))
+    
+    # Guardrail check - scan for forbidden content
+    forbidden_terms = ["legend of zelda", "hyrule", "princess", "nintendo", "triforce"]
+    response_lower = response.content.lower()
+    
+    if any(term in response_lower for term in forbidden_terms):
+        print(f"🚨 Guardrail triggered! Regenerating response...")
+        # Use a fallback response that's safe
+        fallback_response = f"Listen up, space detective! {hint.split('.')[0]}. Now quit bothering me and get back to solving this mystery!"
+        state["current_hint"] = fallback_response
+        print(f"🛡️ Using guardrail fallback response")
+    else:
+        state["current_hint"] = response.content
+        print(f"✅ Guardrail passed - response is clean")
+    
+    print(f"✨ Zelda's version: {state['current_hint'][:100]}{'...' if len(state['current_hint']) > 100 else ''}")
+    print(f"🔍 MAINTAIN_CHARACTER OUTPUT: '{state['current_hint'][:50]}...' -> current_hint updated")
+    return state
