+++ src/verify_correctness_node.py
@@ -1,7 +1,8 @@
-class VerifyCorrectnessNode:
-    def __init__(self, state):
-        self.state = state
+from state import LangGraphState
 
-    def verify_hint(self, hint):
-        # Stub logic for verification, just pass through for now
-        return hint
+def verify_correctness_node(state: LangGraphState) -> LangGraphState:
+    """Stub verification - just pass through for now"""
+    hint = state["current_hint"]
+    print(f"✅ Verifying hint (stub - passing through)")
+    # For now, just pass through
+    return state
